WEBTOON/requirements.txt
flask
requests
firebase-admin
google-generativeai
bs4
gunicorn
